Consolidate metrics emit helpers
Keep the small metric emission helpers, runtime reporter, and their tests together so the metrics package has fewer one-function files.
